requirements.txt
# requirements.txt
# PDF to EPUB Converter Dependencies

# Core Libraries
# fitz==0.0.1.dev2
PyMuPDF==1.25.2
ebooklib==0.18
gradio==4.14.0

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Optional for data validation (future)
# pydantic
# marshmallow